+++ src/DomainSessionException.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace Cadre\Domain_Session;
 
 use Exception;