+++ src/DomainSessionStorageInterface.php
@@ -1,11 +1,11 @@
 <?php
+declare(strict_types=1);
 namespace Cadre\Domain_Session;
 
 interface DomainSessionStorageInterface
 {
-    public function newId();
-    public function read($id);
-    public function write($id, array $data);
-    public function rename($oldId, $newId);
-    public function delete($id);
+    public function createNew($interval = 'PT3M'): DomainSession;
+    public function read(string $id): DomainSession;
+    public function write(DomainSessionInterface $session);
+    public function delete(string $id);
 }